# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Finna FinOps API
  version: 0.2.0
  description: 'The Finna FinOps API is the backend for the Finna cloud cost management platform. It handles multi-cloud cost
    ingestion (Azure Cost Management, GCP Billing, LLM usage), resource wastage detection, alert management, and per-project
    FinOps dashboards. Extracted cost records are normalised into a unified schema and made available through a rich query
    layer with time-window shortcuts (`mtd`, `7d`, `30d`, `90d`), provider filters, and CSV export.


    Authentication supports three flows: username/password JWT (primary), GitHub OAuth (developer login), and pluggable OIDC
    providers with PKCE and auto-provisioning. All resource endpoints require a valid JWT Bearer token. Admin-only endpoints
    (auth provider CRUD) require an `is_admin` claim in the token.


    This repository contains the API-only backend. The React/TypeScript UI lives in a separate repo and communicates exclusively
    through this API.'
  contact:
    name: Finna Team
    email: andrea.carmisciano@abstract.it
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /api/v1/extractors:
    get:
      tags:
      - Extractors
      summary: List extractor registry
      description: Returns registered extractors joined with their associated cloud credential config name. Supports provider
        filtering and pagination.
      operationId: list_extractors_api_v1_extractors_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Page Size
      responses:
        '200':
          description: Extractor list
          content:
            application/json:
              example:
                data:
                - id: e1
                  name: Azure Billing
                  provider: azure
                  config_id: cfg-1
                  config_name: My Azure SP
                  enabled: true
                  schedule: 0 2 * * *
                  last_run: '2026-06-07T02:00:00'
                  status: idle
                count: 1
                total: 1
                page: 1
                page_size: 50
                has_next: false
                has_prev: false
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Extractors
      summary: Register new extractor
      description: Creates a new extractor registry entry linking a provider, a cloud credential config, and a cron schedule.
      operationId: create_extractor_api_v1_extractors_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Data
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Extractor Api V1 Extractors Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/extractors/runs:
    get:
      tags:
      - Extractors
      summary: List extractor runs
      description: Returns recent extractor run records, optionally filtered by provider. Includes status, log output length,
        and timestamps.
      operationId: list_extractor_runs_api_v1_extractors_runs_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Extractor Runs Api V1 Extractors Runs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/extractors/runs/{run_id}:
    get:
      tags:
      - Extractors
      summary: Get extractor run status
      description: Fetches the current status and metadata of a specific run. Returns 404 if the run ID is unknown.
      operationId: get_extractor_run_api_v1_extractors_runs__run_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Extractor Run Api V1 Extractors Runs  Run Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/extractors/runs/{run_id}/cancel:
    post:
      tags:
      - Extractors
      summary: Cancel extractor run
      description: Attempts to cancel an in-progress run. Returns 404 if the run is not found or has already finished.
      operationId: cancel_extractor_run_api_v1_extractors_runs__run_id__cancel_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Cancel Extractor Run Api V1 Extractors Runs  Run Id  Cancel Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/extractors/runs/{run_id}/logs:
    get:
      tags:
      - Extractors
      summary: Get extractor run logs
      description: Returns the sanitised log output for a run as an array of lines. Useful for debugging extraction failures.
      operationId: get_extractor_run_logs_api_v1_extractors_runs__run_id__logs_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Extractor Run Logs Api V1 Extractors Runs  Run Id  Logs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/extractors/run:
    post:
      tags:
      - Extractors
      summary: Trigger extractor run (legacy)
      description: Legacy endpoint that starts an extractor run by `config_id`. Prefer the registry-based flow. Requires `config_id`
        in the request body.
      operationId: run_extractor_via_post_api_v1_extractors_run_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Data
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Run Extractor Via Post Api V1 Extractors Run Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/extractors/{extractor_id}:
    get:
      tags:
      - Extractors
      summary: Get extractor by ID
      description: Fetches a single extractor registry entry. Returns 404 if not found.
      operationId: get_extractor_api_v1_extractors__extractor_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: extractor_id
        in: path
        required: true
        schema:
          type: string
          title: Extractor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Extractor Api V1 Extractors  Extractor Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Extractors
      summary: Delete extractor
      description: Removes the extractor from the registry. In-flight runs are not affected.
      operationId: delete_extractor_api_v1_extractors__extractor_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: extractor_id
        in: path
        required: true
        schema:
          type: string
          title: Extractor Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/token
servers:
- url: http://localhost:8000
  description: Local development server
...